requirements.txt
requests>=2.25.1
ascii-colors>=0.11.5
pillow
pipmaster>=1.1.0
tiktoken
pydantic
numpy
jsonschema
prompt-toolkit>=3.0.0
